refactor(documentation): added demos and refactored docs
