+++ docs/components/box/index.html
@@ -1,10 +1,10 @@
 ---
 title: Box
 also:
+  components/grid: Grid
   components/layouts: Layouts
-  styleguide/grid: Grid
 ---
-{% extends 'styleguide.njk' %}
+{% extends 'component.njk' %}
 {% block content %}
   <section>
     <p>
