+++ docs/components/breadcrumbs/index.html
@@ -3,48 +3,45 @@ title: Breadcrumbs
 ---
 {% extends 'component.njk' %}
 {% block content %}
-<!-- explanation goes here -->
+  <section>
+    <h2 class="hxSectionTitle" id="basic-breadcrumbs">Basic Breadcrumbs</h2>
+    <div class="demo">
+      <nav class="hxBreadcrumb">
+        <a href="#">Home</a>
+      </nav>
 
-<section>
-  <h2 class="hxSectionTitle" id="basic-breadcrumbs">Basic Breadcrumbs</h2>
-  <div class="demo">
-    <nav class="hxBreadcrumb">
-      <a href="#">Home</a>
-    </nav>
+      <nav class="hxBreadcrumb">
+        <a href="#">Home</a>
+        <hx-icon class="delimiter" type="angle-right"></hx-icon>
+        <a href="#">Library</a>
+      </nav>
 
-    <nav class="hxBreadcrumb">
-      <a href="#">Home</a>
-      <hx-icon class="delimiter" type="angle-right"></hx-icon>
-      <a href="#">Library</a>
-    </nav>
+      <nav class="hxBreadcrumb">
+        <a href="#">Home</a>
+        <hx-icon class="delimiter" type="angle-right"></hx-icon>
+        <a href="#">Library</a>
+        <hx-icon class="delimiter" type="angle-right"></hx-icon>
+        <a href="#">Current</a>
+      </nav>
+    </div>
+    {% code 'html' %}
+      <nav class="hxBreadcrumb">
+        <a href="#">Home</a>
+      </nav>
 
-    <nav class="hxBreadcrumb">
-      <a href="#">Home</a>
-      <hx-icon class="delimiter" type="angle-right"></hx-icon>
-      <a href="#">Library</a>
-      <hx-icon class="delimiter" type="angle-right"></hx-icon>
-      <a href="#">Current</a>
-    </nav>
-  </div>
-  {% code 'html' %}
-    <nav class="hxBreadcrumb">
-      <a href="#">Home</a>
-    </nav>
+      <nav class="hxBreadcrumb">
+        <a href="#">Home</a>
+        <hx-icon class="delimiter" type="angle-right"></hx-icon>
+        <a href="#">Library</a>
+      </nav>
 
-    <nav class="hxBreadcrumb">
-      <a href="#">Home</a>
-      <hx-icon class="delimiter" type="angle-right"></hx-icon>
-      <a href="#">Library</a>
-    </nav>
-
-    <nav class="hxBreadcrumb">
-      <a href="#">Home</a>
-      <hx-icon class="delimiter" type="angle-right"></hx-icon>
-      <a href="#">Library</a>
-      <hx-icon class="delimiter" type="angle-right"></hx-icon>
-      <a href="#">Current</a>
-    </nav>
-  {% endcode %}
-</section>
-
+      <nav class="hxBreadcrumb">
+        <a href="#">Home</a>
+        <hx-icon class="delimiter" type="angle-right"></hx-icon>
+        <a href="#">Library</a>
+        <hx-icon class="delimiter" type="angle-right"></hx-icon>
+        <a href="#">Current</a>
+      </nav>
+    {% endcode %}
+  </section>
 {% endblock %}
