+++ docs/components/grid/index.html
@@ -1,10 +1,10 @@
 ---
 title: Grid
 also:
+  components/box: Box
   components/layouts: Layouts
-  styleguide/box: Box
 ---
-{% extends 'styleguide.njk' %}
+{% extends 'component.njk' %}
 {% block content %}
 <section>
   <div class="hxAlert hxAlert--warning">
