+++ docs/components/panels/index.html
@@ -13,21 +13,6 @@ also:
       A panel is a utility for sectioning content vertically into a head,
       body, and foot.
     </p>
-    <ul>
-      <li>
-        The head is fixed to the top of the panel.
-        <ul>
-          <li><em>The head is optional.</em></li>
-        </ul>
-      </li>
-      <li>The body fills the space between the head and foot.</li>
-      <li>
-        The foot is fixed to the bottom of the panel.
-        <ul>
-          <li><em>The foot is optional.</em></li>
-        </ul>
-      </li>
-    </ul>
   </section>
   <section id="vue-panelDemo">
     <h2 class="hxSectionTitle" id="basic-panel">Basic Panel</h2>
