+++ docs/components/popovers/index.html
@@ -1,0 +1,70 @@
+---
+title: Popovers
+also:
+  elements/hx-popover: <hx-popover>
+---
+{% extends 'component.njk' %}
+{% block content %}
+  <section id="vue-popoverDemo">
+    <h2 class="hxSectionTitle" id="tooltip">Basic Popover</h2>
+    <div class="hxRow">
+      <div class="hxCol hxSpan-9-xs">
+        <div class="demo popover-demo">
+          <button data-popover="myPopover" class="hxBtn">Toggle</button>
+
+          <hx-popover id="myPopover"
+            :position="position.value"
+            v-for="_position in positions"
+            v-if="_position.value === position.value">
+
+            <header slot="header">
+              <h4 class="hxContainerTitle">Popover Header</h4>
+            </header>
+            <p>This is my popover body</p>
+            <footer slot="footer">
+              <button class="hxBtn hxBtn--primary">Submit</button>
+              <button class="hxBtn">Cancel</button>
+            </footer>
+          </hx-popover>
+        </div>
+        {% code 'html' %}
+          <button data-popover="myPopover" class="hxBtn">Toggle</button>
+          <hx-popover id="myPopover" position="bottom-right">
+            <header slot="header">
+              <h4 class="hxContainerTitle">Popover Header</h4>
+            </header>
+            <p>This is my popover body</p>
+            <footer slot="footer">
+              <button class="hxBtn hxBtn--primary">Submit</button>
+              <button class="hxBtn">Cancel</button>
+            </footer>
+          </hx-popover>
+        {% endcode %}
+      </div>
+      <div class="hxCol hxSpan-3-xs">
+        Position:
+        <select v-model="position">
+          <option v-for="_position in positions" :value="_position">
+            {% raw %}{{ _position.label }}{% endraw %}
+          </option>
+        </select>
+      </div>
+    </div>
+  </section>
+
+  <section>
+    <h2 class="hxSectionTitle" id="appearance">Appearance</h2>
+    <div class="popover-appearance">
+      {% for _position in [
+        'top', 'top-left', 'top-right',
+        'bottom', 'bottom-left', 'bottom-right',
+        'right', 'right-top', 'right-bottom',
+        'left', 'left-top', 'left-bottom'
+      ] %}
+        <hx-popover position="{{_position}}">
+          <code>position="{{_position}}"</code>
+        </hx-popover>
+      {% endfor %}
+    </div>
+  </section>
+{% endblock %}
