+++ docs/components/reveals/index.html
@@ -1,0 +1,28 @@
+---
+title: Reveals
+also:
+  elements/hx-disclosure: <hx-disclosure>
+  elements/hx-reveal: <hx-reveal>
+---
+{% extends 'component.njk' %}
+{% block content %}
+  <section>
+    <h2 class="hxSectionTitle" id="demo">Demo</h2>
+    <div class="demo">
+      <hx-disclosure class="hxBtn" aria-controls="demo-default-reveal">
+        Click me to show content!
+      </hx-disclosure>
+      <hx-reveal id="demo-default-reveal">
+        <p>You can't see me</p>
+      </hx-reveal>
+    </div>
+    {% code 'html' %}
+      <hx-disclosure class="hxBtn" aria-controls="demo-default-reveal">
+        Click me to show content!
+      </hx-disclosure>
+      <hx-reveal id="demo-default-reveal">
+        <p>You can't see me</p>
+      </hx-reveal>
+    {% endcode %}
+  </section>
+{% endblock %}
