+++ docs/docs.less
@@ -63,20 +63,20 @@ main {
 
 // Component Status Styles
 hx-status {
-  &.-hxFill {
-    &.-hxDeprecated {
+  &.hxFill {
+    &.hxDeprecated {
       background-color: @gray-1000;
       border-color: @gray-900;
       color: @gray-0;
     }
 
-    &.-hxInProgress {
+    &.hxInProgress {
       background-color: @yellow-900;
       border-color: @yellow-500;
       color: @gray-1000;
     }
 
-    &.-hxStable {
+    &.hxStable {
       background-color: @green-900;
       border-color: @green-900;
       color: @gray-0;
