+++ docs/elements/hx-popover/index.html
@@ -1,5 +1,7 @@
 ---
 title: <hx-popover>
+also:
+  components/popovers: Popovers
 ---
 {% extends 'element.njk' %}
 {% block content %}
@@ -20,67 +22,6 @@ title: <hx-popover>
       </hx-def>
     </hx-dl>
   </section>
-
-<section id="vue-popoverDemo">
-  <p>
-    Position:
-    <select v-model="position">
-      <option
-        :value="_position"
-        v-for="_position in positions"
-        v-text="_position.label">
-      </option>
-    <select>
-  </p>
-  <div class="demo popover-demo">
-    <button data-popover="myPopover" class="hxBtn">Toggle</button>
-
-    <hx-popover id="myPopover"
-      :position="position.value"
-      v-for="_position in positions"
-      v-if="_position.value === position.value">
-
-      <header slot="header">
-        <h4 class="hxContainerTitle">Popover Header</h4>
-      </header>
-      <p>This is my popover body</p>
-      <footer slot="footer">
-        <button class="hxBtn hxBtn--primary">Submit</button>
-        <button class="hxBtn">Cancel</button>
-      </footer>
-    </hx-popover>
-  </div>
-  {% code 'html' %}
-    <button data-popover="myPopover" class="hxBtn">Toggle</button>
-
-    <hx-popover id="myPopover" position="bottom-right">
-      <header slot="header">
-        <h4 class="hxContainerTitle">Popover Header</h4>
-      </header>
-      <p>This is my popover body</p>
-      <footer slot="footer">
-        <button class="hxBtn hxBtn--primary">Submit</button>
-        <button class="hxBtn">Cancel</button>
-      </footer>
-    </hx-popover>
-  {% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="appearance">Appearance</h2>
-  <div class="popover-appearance">
-    {% for _position in [
-      'top', 'top-left', 'top-right',
-      'bottom', 'bottom-left', 'bottom-right',
-      'right', 'right-top', 'right-bottom',
-      'left', 'left-top', 'left-bottom'
-    ] %}
-      <hx-popover position="{{_position}}">
-        <code>position="{{_position}}"</code>
-      </hx-popover>
-    {% endfor %}
-  </div>
-</section>
 {% endblock %}
 
 {% block attributes %}
