+++ docs/elements/hx-status/index.html
@@ -1,5 +1,7 @@
 ---
 title: <hx-status>
+also:
+  components/status-pills: Status Pills
 ---
 {% extends 'element.njk' %}
 {% block content %}
@@ -15,110 +17,4 @@ title: <hx-status>
       </hx-def>
     </hx-dl>
   </section>
-
-<section>
-  <h2 class="hxSectionTitle" id="default-label">Default Label</h2>
-  <div class="demo">
-    <hx-status>
-      Default Label
-    </hx-status>
-  </div>
-  {% code 'html' %}
-    <hx-status>
-      Default Label
-    </hx-status>
-  {% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="emphasis-gray">Emphasis Gray</h2>
-  <p>
-    The <code>.-hxEmphasisGray</code> CSS class can be added to status labels that require more
-    attention than the default style.
-  </p>
-  <div class="demo">
-    <hx-status class="-hxEmphasisGray">
-      Emphasis Gray
-    </hx-status>
-  </div>
-  {% code 'html' %}
-    <hx-status class="-hxEmphasisGray">
-      Emphasis Gray
-    </hx-status>
-  {% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="emphasis-purple">Emphasis Purple</h2>
-  <p>
-    The <code>.-hxEmphasisPurple</code> CSS class can be added to status labels that require
-    more attention than the gray emphasis.
-  </p>
-  <div class="demo">
-    <hx-status class="-hxEmphasisPurple">
-      Emphasis Purple
-    </hx-status>
-  </div>
-  {% code 'html' %}
-    <hx-status class="-hxEmphasisPurple">
-      Emphasis Purple
-    </hx-status>
-  {% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="subdued">Subdued</h2>
-  <p>
-    The <code>.-hxSubdued</code> class can be used on status labels that require less or
-    no attention from the user.
-  </p>
-  <div class="demo">
-    <hx-status class="-hxSubdued">
-      Subdued
-    </hx-status>
-  </div>
-  {% code 'html' %}
-    <hx-status class="-hxSubdued">
-      Subdued
-    </hx-status>
-  {% endcode %}
-</section>
-
-<section>
-  <h2 class="hxSectionTitle" id="filled-labels">Filled Labels</h2>
-  <p>
-    The <code>.-hxFill</code> class can be used on status labels to apply filled
-    backgrounds. This class can be used in conjunction with any of the previous
-    status label modifier classes above.
-  </p>
-  <div class="demo">
-    <hx-status class="-hxFill">
-      Default Label
-    </hx-status><br/>
-    <hx-status class="-hxFill -hxEmphasisGray">
-      Emphasis Gray
-    </hx-status><br/>
-    <hx-status class="-hxFill -hxEmphasisPurple">
-      Emphasis Purple
-    </hx-status><br/>
-    <hx-status class="-hxFill -hxSubdued">
-      Subdued
-    </hx-status>
-  </div>
-  {% code 'html' %}
-    <hx-status class="-hxFill">
-      Default Label
-    </hx-status><br/>
-    <hx-status class="-hxFill -hxEmphasisGray">
-      Emphasis Gray
-    </hx-status><br/>
-    <hx-status class="-hxFill -hxEmphasisPurple">
-      Emphasis Purple
-    </hx-status><br/>
-    <hx-status class="-hxFill -hxSubdued">
-      Subdued
-    </hx-status>
-  {% endcode %}
-</section>
-
 {% endblock %}
