+++ docs/styles/demo/icon-docs.less
@@ -1,5 +1,4 @@
 .icon-demo {
-  height: 100%;
   text-align: center;
 
   & > p {
