+++ src/helix-ui.less
@@ -102,6 +102,7 @@ fieldset label {
 @import 'core/hx-tablist';
 @import 'core/hx-tabpanel';
 @import 'core/hx-tabset';
+@import 'core/hx-tooltip';
 /* ====== end:BASE ========== */
