+++ src/helix-ui/styles/core/hx-status.less
@@ -15,39 +15,39 @@ hx-status {
   text-overflow: ellipsis;
   white-space: nowrap;
 
-  &.-hxEmphasisGray {
+  &.hxEmphasisGray {
     border-color: @gray-900;
     color: @gray-900;
   }
 
-  &.-hxEmphasisPurple {
+  &.hxEmphasisPurple {
     border-color: @purple-500;
     color: @purple-500;
   }
 
-  &.-hxSubdued {
+  &.hxSubdued {
     border-color: @gray-600;
     color: @gray-600;
   }
 
-  &.-hxFill {
+  &.hxFill {
     background-color: @gray-750;
     border-color: @gray-750;
     color: @gray-0;
 
-    &.-hxEmphasisGray {
+    &.hxEmphasisGray {
       background-color: @gray-900;
       border-color: @gray-900;
       color: @gray-0;
     }
 
-    &.-hxEmphasisPurple {
+    &.hxEmphasisPurple {
       background-color: @purple-500;
       border-color: @purple-500;
       color: @gray-0;
     }
 
-    &.-hxSubdued {
+    &.hxSubdued {
       background-color: @gray-600;
       border-color: @gray-600;
       color: @gray-0;
