+++ src/helix-ui/styles/core/hx-tooltip.less
@@ -1,0 +1,31 @@
+hx-tooltip {
+    background-color: @gray-0;
+    border-color: @gray-300;
+    border-style: solid;
+    border-width: 1px;
+    box-shadow: 0px 3px 3px 0 @shadow-color;
+    color: initial;
+    display: none;
+    font-family: "Roboto", sans-serif;
+    font-size: 1rem;
+    font-style: normal;
+    font-weight: normal;
+    left: 0;
+    letter-spacing: normal;
+    max-width: 25rem;
+    position: absolute;
+    text-align: left;
+    text-decoration: none;
+    text-shadow: none;
+    text-transform: none;
+    top: 0;
+    white-space: normal;
+    word-break: normal;
+    word-spacing: normal;
+    word-wrap: normal;
+    z-index: @tooltip-z-index;
+
+    &[open] {
+        display: block;
+    }
+}
