+++ src/helix-ui/styles/modules/text-input.less
@@ -37,7 +37,7 @@
     outline: 0;
     border-color: @cyan-700;
     box-shadow: @focus-glow;
-    &.-hxInvalid {
+    &[invalid] {
       box-shadow: @focus-glow-invalid;
     }
   }
